First Release (1.0)
Features:
- AI based zombie following, aka
   pathfinding.
- Health and damage systems in place
   for both zombies and players.
- Smooth camera following.
- Artificial 3D effect when throwing the
   ball, to give the illusion of a three--
   dimensional game.
- Limited ammunition.
- Ball can be picked up by player and
   reused.
Bug fixes:
- Ball projectile wall clipping.
- Scene disappearing and every were
   just gone as soon as you first moved.

The last thing missing is a the textures and the play area to walk around in, but the groundworks for making that possible is already in place, meaning as soon as we have the textures and a world layout, it should take no more then ten minutes to be with a finished product.
Of course you can always add Quality of Life features, which is most likely what I will be doing tomorrow.